requirements.txt
# Core dependencies
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
python-dotenv>=1.0.0
loguru>=0.7.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# pytest-mock>=3.10.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.0.0
# pre-commit>=3.0.0
